requirements.txt
# RAG Logistics Assistant Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
